chore: modularize engine registration with build tags to favour custom builds
