+++ backend/pkg/cryptoengines/builder/builder.go
@@ -7,13 +7,7 @@ import (
 	"github.com/lamassuiot/lamassuiot/core/v3/pkg/engines/cryptoengines"
 	log "github.com/sirupsen/logrus"
 
-	"github.com/lamassuiot/lamassuiot/engines/crypto/aws/v3"
-	aws_subsystem "github.com/lamassuiot/lamassuiot/engines/crypto/aws/v3/subsystem"
 	"github.com/lamassuiot/lamassuiot/engines/crypto/filesystem/v3"
-	"github.com/lamassuiot/lamassuiot/engines/crypto/pkcs11/v3"
-	pkcs11_subsystem "github.com/lamassuiot/lamassuiot/engines/crypto/pkcs11/v3/subsystem"
-	"github.com/lamassuiot/lamassuiot/engines/crypto/vaultkv2/v3"
-	vault_subsystem "github.com/lamassuiot/lamassuiot/engines/crypto/vaultkv2/v3/subsystem"
 )
 
 func BuildCryptoEngine(logger *log.Entry, conf cconfig.CryptoEngineConfig) (cryptoengines.CryptoEngine, error) {
@@ -26,16 +20,6 @@ func BuildCryptoEngine(logger *log.Entry, conf cconfig.CryptoEngineConfig) (cryp
 }
 
 func init() {
-	log.Info("Registering crypto engines")
+	log.Info("Registering default crypto engine")
 	filesystem.Register()
-
-	aws.RegisterAWSKMS()
-	aws.RegisterAWSSecrets()
-	aws_subsystem.Register()
-
-	vaultkv2.Register()
-	vault_subsystem.Register()
-
-	pkcs11.Register()
-	pkcs11_subsystem.Register()
 }
