+++ backend/pkg/cryptoengines/builder/registrar_pkcs11.go
@@ -1,0 +1,15 @@
+//go:build !nopkcs11
+
+package builder
+
+import (
+	"github.com/lamassuiot/lamassuiot/engines/crypto/pkcs11/v3"
+	pkcs11_subsystem "github.com/lamassuiot/lamassuiot/engines/crypto/pkcs11/v3/subsystem"
+	log "github.com/sirupsen/logrus"
+)
+
+func init() {
+	log.Info("Registering VaultKV crypto engines")
+	pkcs11.Register()
+	pkcs11_subsystem.Register()
+}
