+++ backend/pkg/eventbus/builder/registrar_aws.go
@@ -1,0 +1,9 @@
+//go:build !noaws
+
+package builder
+
+import "github.com/lamassuiot/lamassuiot/engines/eventbus/aws/v3"
+
+func init() {
+	aws.Register()
+}
